Revert "Finished enemy and puzzle code"

+++ Assets/Scripts/HealthUI.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
@@ -11,7 +10,7 @@ public class HealthUI : MonoBehaviour
 
     private List<Image> hearts = new List<Image>();
 
-    public void SetMaxHearts(float maxHearts)
+    public void SetMaxHearts(int maxHearts)
     {
         foreach (Image heart in hearts)
         {
@@ -31,7 +30,7 @@ public class HealthUI : MonoBehaviour
         Debug.Log("Created hearts: " + hearts.Count);
     }
 
-    public void UpdateHearts(float currentHealth)
+    public void UpdateHearts(int currentHealth)
     {
         Debug.Log("Updating hearts to: " + currentHealth);
 
@@ -49,5 +48,4 @@ public class HealthUI : MonoBehaviour
             }
         }
     }
-
 }
